projects/Doosan_Bob/products_cloudstorage/requirements.txt
attrs==19.3.0
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
configparser==5.0.0
cycler==0.10.0
dask==2.22.0
datalab==1.2.0
DateTime==4.3
decorator==4.4.2
fsspec==0.8.0
future==0.18.2
gcsfs==0.6.2
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.26.1
google-cloud-bigquery-storage==1.0.0
google-cloud-core==1.4.1
google-cloud-monitoring==0.31.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpcio==1.31.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
kiwisolver==1.2.0
locket==0.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mock==4.0.2
networkx==2.4
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.1.0
pandas-gbq==0.13.2
pandas-profiling==1.4.0
parso==0.7.1
partd==1.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipedrive-python-lib==1.1.4
plotly==4.9.0
prompt-toolkit==3.0.6
protobuf==3.12.4
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.1.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.7.24
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
zope.interface==5.1.0
